Update `swap_lon_axis()` to drop extra cell to handle prime meridian
