+++ xcdat/bounds.py
@@ -1003,51 +1003,3 @@ def create_bounds(axis: CFAxisKey, coord_var: xr.DataArray) -> xr.DataArray:
     )
 
     return bounds
-
-
-def _get_bounds_dim(da_coords: xr.DataArray, da_bounds: xr.DataArray) -> str:
-    """Identify the bounds dimension in the given bounds DataArray.
-
-    This function determines which dimension in the bounds DataArray
-    represents the bounds (e.g., lower and upper limits) for each value
-    in the coordinate DataArray. According to the CF Conventions, a boundary
-    variable will have one more dimension than its associated coordinate or
-    auxiliary coordinate variable. Refer to [6] for more details.
-
-    Parameters
-    ----------
-    da_coords : xr.DataArray
-        The coordinate DataArray.
-    da_bounds : xr.DataArray
-        The bounds DataArray.
-
-    Returns
-    -------
-    str
-        The name of the bounds dimension.
-
-    Raises
-    ------
-    ValueError
-        If the bounds variable does not have exactly one more dimension than
-        the coordinate variable, or if no valid bounds dimension is found.
-
-    References
-    ----------
-    .. [6] https://cfconventions.org/Data/cf-conventions/cf-conventions-1.10/cf-conventions.html#cell-boundaries
-    """
-    bounds_dim = [dim for dim in da_bounds.dims if dim not in da_coords.dims]
-
-    if len(bounds_dim) == 1:
-        return str(bounds_dim[0])
-    elif len(bounds_dim) == 0:
-        raise ValueError(
-            "No extra dimension found in bounds variable compared to coordinate "
-            f"variable. Coordinate dims: {da_coords.dims}, bounds dims: "
-            f"{da_bounds.dims}"
-        )
-    else:
-        raise ValueError(
-            f"Bounds variable must have exactly one more dimension than the coordinate "
-            f"variable. Coordinate dims: {da_coords.dims}, bounds dims: {da_bounds.dims}"
-        )
